requirements.txt
# Model Router Requirements
# Install with: pip install -r requirements.txt

# Core Flask and API
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.9.0

# AI Model SDKs
openai>=1.50.0
anthropic>=0.39.0
xai-sdk>=0.1.0  # xAI Grok API
google-generativeai>=0.8.0  # Google Gemini 2.5 Pro API

# Environment and utilities
python-dotenv>=1.0.0

# For WebSocket support
eventlet>=0.33.3

# Optional but recommended
requests>=2.31.0  # For testing API calls
colorama>=0.4.6   # For colored terminal output